pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py4lexis"
version = "1.0.0"
description = "Provides functions to manage datasets within LEXIS platform."
readme = "README.md"
authors = [{ name = "Martin Golasowski", email = "martin.golasowski@vsb.cz" },
           { name = "Ladislav Foltyn", email = "ladislav.foltyn@vsb.cz" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
keywords = ["hpc", "py4lexis", "lexis"]
dependencies = [
    "tabulate == 0.8.10",
    "requests == 2.27.1",
    "python-keycloak == 0.27.0",
    "python-irodsclient == 1.1.3",
    "tomli == 2.0.1",
    "tuspy == 0.2.5"
]
requires-python = ">=3.7"

    [project.optional-dependencies]
    build = ["build", "twine"]
    dev   = ["black", "bumpver", "pytest"]

